BSBD-7.1 - Restructure and migrate to v1 directory
